GramSetu/grampanchayat-app/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
PyMySQL==1.1.1
gunicorn==21.2.0
